% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-header.R
\name{.bullets}
\alias{.bullets}
\title{Build a \code{cli::cli_bullets()}-ready named character vector}
\usage{
.bullets(symbol, messages)
}
\description{
Build a \code{cli::cli_bullets()}-ready named character vector
}
\keyword{internal}
