% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-helper.R
\name{.header_field_set}
\alias{.header_field_set}
\title{Assign a value at a (possibly nested) dotted path into a list}
\usage{
.header_field_set(fm, path, value)
}
\description{
Assign a value at a (possibly nested) dotted path into a list
}
\keyword{internal}
