% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header-helper.R
\name{.header_helper_ask_yn}
\alias{.header_helper_ask_yn}
\title{Ask a yes/no question}
\usage{
.header_helper_ask_yn(prompt, default = FALSE)
}
\arguments{
\item{prompt}{Question text (without a yes/no hint -- one is appended).}

\item{default}{Default answer if the user just presses enter.}
}
\value{
\code{TRUE}/\code{FALSE}.
}
\description{
Ask a yes/no question
}
\keyword{internal}
