% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-header.R
\name{.print_header_validation}
\alias{.print_header_validation}
\title{cli-formatted report for \code{validate_header()}}
\usage{
.print_header_validation(
  shell_qmd,
  provided,
  missing_required,
  missing_recommended,
  available,
  problems
)
}
\description{
cli-formatted report for \code{validate_header()}
}
\keyword{internal}
